[25127] Don't reload list if query was initialized before
When moving from show to list, the list controller is reinitialized
and the list is reloaded through `loadQuery()`.

We can bypass this by simply checking whether the query has ever been
checked by the checksum service. Since the query observers are still
registered, changes to the query are still caught.